test: updated test cases to meet the new config group for lint
