+++ tests/dev_config.rs
@@ -517,3 +517,103 @@ config :phoenix, :json_library, Jason
 
     teardown(wk_temp, elixir_temp)
 }
+
+#[test]
+#[serial]
+fn test_wukong_dev_config_lint_success() {
+    let wk_temp = assert_fs::TempDir::new().unwrap();
+    let main_config_file = wk_temp.child("config/config.exs");
+    main_config_file.touch().unwrap();
+
+    main_config_file
+        .write_str(
+            r#"
+use Mix.Config
+
+System.get_env("API_KEY")
+System.fetch_env("API_SECRET")
+System.fetch_env!("API_TOKEN")
+
+# invalid
+import_config "config/dev.exs"
+
+# valid
+if File.exists?("config/dev.exs") do
+  import_config "config/dev.exs"
+end
+
+# invalid
+if File.exists?("config/a.exs") do
+  import_config "config/b.exs"
+end
+
+# valid
+File.exists?("config/dev.exs") && import_config "config/dev.exs"
+# invalid
+File.exists?("config/a.exs") && import_config "config/b.exs"
+
+test_domain = System.get_env("TEST_DOMAIN", "mv.test.com")
+
+# Use Jason for JSON parsing in Phoenix
+config :phoenix, :json_library, Jason
+    "#,
+        )
+        .unwrap();
+
+    let dev_config_file = wk_temp.child("config/dev.exs");
+    dev_config_file.touch().unwrap();
+
+    dev_config_file
+        .write_str(
+            r#"
+use Mix.Config
+
+System.get_env("API_KEY")
+System.fetch_env("API_SECRET")
+System.fetch_env!("API_TOKEN")
+
+# invalid
+import_config "config/dev.exs"
+
+# valid
+if File.exists?("config/dev.exs") do
+  import_config "config/dev.exs"
+end
+
+# invalid
+if File.exists?("config/a.exs") do
+  import_config "config/b.exs"
+end
+
+# valid
+File.exists?("config/dev.exs") && import_config "config/dev.exs"
+# invalid
+File.exists?("config/a.exs") && import_config "config/b.exs"
+
+test_domain = System.get_env("TEST_DOMAIN", "mv.test.com")
+
+# Use Jason for JSON parsing in Phoenix
+config :phoenix, :json_library, Jason
+    "#,
+        )
+        .unwrap();
+
+    let cmd = common::wukong_raw_command()
+        .arg("dev")
+        .arg("config")
+        .arg("lint")
+        .arg(wk_temp.path().to_str().unwrap())
+        .assert()
+        .failure();
+
+    let output = cmd.get_output();
+
+    insta::with_settings!({filters => vec![
+        (format!("{}", wk_temp.path().to_str().unwrap()).as_str(), "[TEMP_DIR]"),
+        (r"\d+(.\d+)ms|\d+(.\d+)s", "[DURATION]"),
+    ]}, {
+        insta::assert_snapshot!(std::str::from_utf8(&output.stderr).unwrap());
+    });
+
+    wk_temp.close().unwrap();
+}
